+++ tasks.rb
@@ -216,6 +216,7 @@ def create_hpxmls
     'base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-acca.xml' => 'base-hvac-air-to-air-heat-pump-1-speed.xml',
     'base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-hers.xml' => 'base-hvac-air-to-air-heat-pump-1-speed.xml',
     'base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-maxload.xml' => 'base-hvac-air-to-air-heat-pump-1-speed.xml',
+    'base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-maxload-miami-fl.xml' => 'base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-maxload.xml',
     'base-hvac-autosize-air-to-air-heat-pump-2-speed-sizing-methodology-acca.xml' => 'base-hvac-air-to-air-heat-pump-2-speed.xml',
     'base-hvac-autosize-air-to-air-heat-pump-2-speed-sizing-methodology-hers.xml' => 'base-hvac-air-to-air-heat-pump-2-speed.xml',
     'base-hvac-autosize-air-to-air-heat-pump-2-speed-sizing-methodology-maxload.xml' => 'base-hvac-air-to-air-heat-pump-2-speed.xml',
@@ -478,9 +479,12 @@ def create_hpxmls
           # Make all IDs unique so the HPXML is valid
           new_building_element.each_node do |node|
             next unless node.is_a?(Oga::XML::Element)
-            next if XMLHelper.get_attribute_value(node, 'id').nil?
 
-            XMLHelper.add_attribute(node, 'id', "#{XMLHelper.get_attribute_value(node, 'id')}_#{i}")
+            if not XMLHelper.get_attribute_value(node, 'id').nil?
+              XMLHelper.add_attribute(node, 'id', "#{XMLHelper.get_attribute_value(node, 'id')}_#{i}")
+            elsif not XMLHelper.get_attribute_value(node, 'idref').nil?
+              XMLHelper.add_attribute(node, 'idref', "#{XMLHelper.get_attribute_value(node, 'idref')}_#{i}")
+            end
           end
 
           hpxml_element.children << new_building_element
@@ -2048,6 +2052,10 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['site_state_code'] = 'FL'
     args['weather_station_epw_filepath'] = 'USA_FL_Miami.Intl.AP.722020_TMY3.epw'
     args['heating_system_heating_capacity'] = 12000.0
+  elsif ['base-hvac-autosize-air-to-air-heat-pump-1-speed-sizing-methodology-maxload-miami-fl.xml'].include? hpxml_file
+    args['site_iecc_zone'] = '1A'
+    args['site_state_code'] = 'FL'
+    args['weather_station_epw_filepath'] = 'USA_FL_Miami.Intl.AP.722020_TMY3.epw'
   elsif ['base-location-phoenix-az.xml'].include? hpxml_file
     args['site_iecc_zone'] = '2B'
     args['site_state_code'] = 'AZ'
@@ -2482,7 +2490,7 @@ def apply_hpxml_modification_ashrae_140(hpxml)
   # ------------ #
 
   hpxml.header.xml_generated_by = 'tasks.rb'
-  hpxml.header.created_date_and_time = Time.new(2000, 1, 1).strftime('%Y-%m-%dT%H:%M:%S%:z') # Hard-code to prevent diffs
+  hpxml.header.created_date_and_time = Time.new(2000, 1, 1, 0, 0, 0, '-07:00').strftime('%Y-%m-%dT%H:%M:%S%:z') # Hard-code to prevent diffs
   hpxml.header.apply_ashrae140_assumptions = true
 
   # --------------------- #
@@ -2526,11 +2534,11 @@ def apply_hpxml_modification_ashrae_140(hpxml)
       end
     end
   end
-  hpxml.frame_floors.each do |frame_floor|
-    next unless frame_floor.is_ceiling
+  hpxml.floors.each do |floor|
+    next unless floor.is_ceiling
 
-    frame_floor.interior_finish_type = HPXML::InteriorFinishGypsumBoard
-    frame_floor.interior_finish_thickness = 0.5
+    floor.interior_finish_type = HPXML::InteriorFinishGypsumBoard
+    floor.interior_finish_thickness = 0.5
   end
   hpxml.foundation_walls.each do |fwall|
     if fwall.insulation_interior_r_value == 0
@@ -2591,7 +2599,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
 
   # General logic for all files
   hpxml.header.xml_generated_by = 'tasks.rb'
-  hpxml.header.created_date_and_time = Time.new(2000, 1, 1).strftime('%Y-%m-%dT%H:%M:%S%:z') # Hard-code to prevent diffs
+  hpxml.header.created_date_and_time = Time.new(2000, 1, 1, 0, 0, 0, '-07:00').strftime('%Y-%m-%dT%H:%M:%S%:z') # Hard-code to prevent diffs
 
   # Logic that can only be applied based on the file name
   if ['base-hvac-undersized-allow-increased-fixed-capacities.xml'].include? hpxml_file
@@ -2658,7 +2666,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
 
     roof.interior_finish_type = HPXML::InteriorFinishGypsumBoard
   end
-  (hpxml.walls + hpxml.foundation_walls + hpxml.frame_floors).each do |surface|
+  (hpxml.walls + hpxml.foundation_walls + hpxml.floors).each do |surface|
     if surface.is_a?(HPXML::FoundationWall) && surface.interior_adjacent_to != HPXML::LocationBasementConditioned
       surface.interior_finish_type = HPXML::InteriorFinishNone
     end
@@ -2671,7 +2679,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                  HPXML::LocationAtticVented,
                  HPXML::LocationOtherHousingUnit,
                  HPXML::LocationBasementConditioned].include?(surface.exterior_adjacent_to)
-    next if surface.is_a?(HPXML::FrameFloor) && surface.is_floor
+    next if surface.is_a?(HPXML::Floor) && surface.is_floor
 
     surface.interior_finish_type = HPXML::InteriorFinishGypsumBoard
   end
@@ -2713,12 +2721,12 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit
            }           [0]
     wall.exterior_adjacent_to = adjacent_to
-    hpxml.frame_floors[0].exterior_adjacent_to = adjacent_to
-    hpxml.frame_floors[1].exterior_adjacent_to = adjacent_to
+    hpxml.floors[0].exterior_adjacent_to = adjacent_to
+    hpxml.floors[1].exterior_adjacent_to = adjacent_to
     if hpxml_file != 'base-bldgtype-multifamily-adjacent-to-other-housing-unit.xml'
       wall.insulation_assembly_r_value = 23
-      hpxml.frame_floors[0].insulation_assembly_r_value = 18.7
-      hpxml.frame_floors[1].insulation_assembly_r_value = 18.7
+      hpxml.floors[0].insulation_assembly_r_value = 18.7
+      hpxml.floors[1].insulation_assembly_r_value = 18.7
     end
     hpxml.windows.each do |window|
       window.area *= 0.35
@@ -2781,25 +2789,25 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                     emittance: 0.92,
                     interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 4.0)
-    hpxml.frame_floors[0].delete
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationOtherNonFreezingSpace,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 550,
-                           insulation_assembly_r_value: 18.7,
-                           other_space_above_or_below: HPXML::FrameFloorOtherSpaceBelow)
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 200,
-                           insulation_assembly_r_value: 18.7,
-                           other_space_above_or_below: HPXML::FrameFloorOtherSpaceBelow)
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationOtherHeatedSpace,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 150,
-                           insulation_assembly_r_value: 2.1,
-                           other_space_above_or_below: HPXML::FrameFloorOtherSpaceBelow)
+    hpxml.floors[0].delete
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationOtherNonFreezingSpace,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 550,
+                     insulation_assembly_r_value: 18.7,
+                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 200,
+                     insulation_assembly_r_value: 18.7,
+                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationOtherHeatedSpace,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 150,
+                     insulation_assembly_r_value: 2.1,
+                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
     wall = hpxml.walls.select { |w|
              w.interior_adjacent_to == HPXML::LocationLivingSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherMultifamilyBufferSpace
@@ -2902,12 +2910,12 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.foundation_walls.each do |foundation_wall|
       foundation_wall.area = 1200.0 / hpxml.foundation_walls.size
     end
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationAtticUnvented,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 450,
-                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
-                           insulation_assembly_r_value: 39.3)
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationAtticUnvented,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 450,
+                     interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                     insulation_assembly_r_value: 39.3)
     hpxml.slabs[0].area = 1350
     hpxml.slabs[0].exposed_perimeter = 150
     hpxml.windows[1].area = 108
@@ -3146,12 +3154,12 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                                depth_below_grade: 3,
                                insulation_interior_r_value: 0,
                                insulation_exterior_r_value: 0)
-    hpxml.frame_floors[0].area = 675
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationCrawlspaceUnvented,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 675,
-                           insulation_assembly_r_value: 18.7)
+    hpxml.floors[0].area = 675
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationCrawlspaceUnvented,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 675,
+                     insulation_assembly_r_value: 18.7)
     hpxml.slabs[0].area = 675
     hpxml.slabs[0].exposed_perimeter = 75
     hpxml.slabs.add(id: "Slab#{hpxml.slabs.size + 1}",
@@ -3288,11 +3296,11 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                     solar_absorptance: 0.7,
                     emittance: 0.92,
                     insulation_assembly_r_value: 4)
-    hpxml.frame_floors.add(id: "FrameFloor#{hpxml.frame_floors.size + 1}",
-                           exterior_adjacent_to: HPXML::LocationGarage,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
-                           area: 400,
-                           insulation_assembly_r_value: 39.3)
+    hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
+                     exterior_adjacent_to: HPXML::LocationGarage,
+                     interior_adjacent_to: HPXML::LocationLivingSpace,
+                     area: 400,
+                     insulation_assembly_r_value: 39.3)
     hpxml.slabs[0].area -= 400
     hpxml.slabs[0].exposed_perimeter -= 40
     hpxml.slabs.add(id: "Slab#{hpxml.slabs.size + 1}",
@@ -3518,19 +3526,19 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.foundation_walls << hpxml.foundation_walls[-1].dup
     hpxml.foundation_walls[-1].id += '_tiny'
     hpxml.foundation_walls[-1].area = 0.05
-    for n in 1..hpxml.frame_floors.size
-      hpxml.frame_floors[n - 1].area /= 9.0
+    for n in 1..hpxml.floors.size
+      hpxml.floors[n - 1].area /= 9.0
       for i in 2..9
-        hpxml.frame_floors << hpxml.frame_floors[n - 1].dup
-        hpxml.frame_floors[-1].id += "_#{i}"
+        hpxml.floors << hpxml.floors[n - 1].dup
+        hpxml.floors[-1].id += "_#{i}"
         if hpxml_file == 'base-enclosure-split-surfaces2.xml'
-          hpxml.frame_floors[-1].insulation_assembly_r_value += 0.01 * i
+          hpxml.floors[-1].insulation_assembly_r_value += 0.01 * i
         end
       end
     end
-    hpxml.frame_floors << hpxml.frame_floors[-1].dup
-    hpxml.frame_floors[-1].id += '_tiny'
-    hpxml.frame_floors[-1].area = 0.05
+    hpxml.floors << hpxml.floors[-1].dup
+    hpxml.floors[-1].id += '_tiny'
+    hpxml.floors[-1].area = 0.05
     for n in 1..hpxml.slabs.size
       hpxml.slabs[n - 1].area /= 9.0
       hpxml.slabs[n - 1].exposed_perimeter /= 9.0
@@ -4525,7 +4533,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
      hpxml.rim_joists +
      hpxml.walls +
      hpxml.foundation_walls +
-     hpxml.frame_floors +
+     hpxml.floors +
      hpxml.slabs +
      hpxml.windows +
      hpxml.skylights +
@@ -4543,7 +4551,7 @@ def renumber_hpxml_ids(hpxml)
   { hpxml.walls => 'Wall',
     hpxml.foundation_walls => 'FoundationWall',
     hpxml.rim_joists => 'RimJoist',
-    hpxml.frame_floors => 'FrameFloor',
+    hpxml.floors => 'Floor',
     hpxml.roofs => 'Roof',
     hpxml.slabs => 'Slab',
     hpxml.windows => 'Window',
@@ -4560,8 +4568,8 @@ def renumber_hpxml_ids(hpxml)
         if attic_or_fnd.respond_to?(:attached_to_rim_joist_idrefs) && !attic_or_fnd.attached_to_rim_joist_idrefs.nil? && !attic_or_fnd.attached_to_rim_joist_idrefs.delete(surf.id).nil?
           attic_or_fnd.attached_to_rim_joist_idrefs << "#{surf_name}#{i + 1}"
         end
-        if attic_or_fnd.respond_to?(:attached_to_frame_floor_idrefs) && !attic_or_fnd.attached_to_frame_floor_idrefs.nil? && !attic_or_fnd.attached_to_frame_floor_idrefs.delete(surf.id).nil?
-          attic_or_fnd.attached_to_frame_floor_idrefs << "#{surf_name}#{i + 1}"
+        if attic_or_fnd.respond_to?(:attached_to_floor_idrefs) && !attic_or_fnd.attached_to_floor_idrefs.nil? && !attic_or_fnd.attached_to_floor_idrefs.delete(surf.id).nil?
+          attic_or_fnd.attached_to_floor_idrefs << "#{surf_name}#{i + 1}"
         end
         if attic_or_fnd.respond_to?(:attached_to_slab_idrefs) && !attic_or_fnd.attached_to_slab_idrefs.nil? && !attic_or_fnd.attached_to_slab_idrefs.delete(surf.id).nil?
           attic_or_fnd.attached_to_slab_idrefs << "#{surf_name}#{i + 1}"
@@ -4846,7 +4854,7 @@ def create_schematron_hpxml_validator(hpxml_docs)
   id_names = ['SystemIdentifier',
               'BuildingID']
   idref_names = ['AttachedToRoof',
-                 'AttachedToFrameFloor',
+                 'AttachedToFloor',
                  'AttachedToSlab',
                  'AttachedToFoundationWall',
                  'AttachedToWall',